factor out raw socket adapter
